+++ src/main/java/com/ewan/dunjeon/world/Interactable.java
@@ -11,7 +11,7 @@ public interface Interactable {
 
     public void onInteract(Entity interactor, InteractionType type);
 //    public boolean isInteractable(Entity interactor);
-    public float getPosX();
-    public float getPosY();
+    public double getPosX();
+    public double getPosY();
     public Set<InteractionType> getAvailableInteractions(Entity interactor);
 }
