+++ src/main/java/com/ewan/dunjeon/world/entities/creatures/ChaseAI.java
@@ -17,7 +17,7 @@ public class ChaseAI extends AIState {
 
 
     List<Point> currentPath = new ArrayList<>();
-    float desiredDistToTarget = 1;
+    double desiredDistToTarget = 1;
     long targetUUID;
     boolean targetUnreachable = false;
 
@@ -40,9 +40,9 @@ public class ChaseAI extends AIState {
         }
         if(!targetMemory.isOldData() && cellMemories.stream().noneMatch(cellMemory -> cellMemory.enterable == CellMemory.EnterableStatus.CLOSED)){
 
-            float angle = (float)Math.atan2(targetMemory.getStateData().getY() - hostEntity.getPosY(),targetMemory.getStateData().getX() - hostEntity.getPosX());
+            double angle = (double)Math.atan2(targetMemory.getStateData().getY() - hostEntity.getPosY(),targetMemory.getStateData().getX() - hostEntity.getPosX());
 
-            hostEntity.addVelocity((float)Math.cos(angle) * hostEntity.getWalkSpeed(), (float)Math.sin(angle) * hostEntity.getWalkSpeed());
+            hostEntity.addVelocity((double)Math.cos(angle) * hostEntity.getWalkSpeed(), (double)Math.sin(angle) * hostEntity.getWalkSpeed());
 
         }else {
 
@@ -61,7 +61,7 @@ public class ChaseAI extends AIState {
 
             if (!pathValid) {
                 Point p = new Point((int) Math.floor(targetMemory.getStateData().getX()), (int) Math.floor(targetMemory.getStateData().getY()));
-                float[][] weights = getWeightMapFromMemory();
+                double[][] weights = getWeightMapFromMemory();
                 currentPath = PathFinding.getAStarPath(weights, new Point((int) hostEntity.getPosX(), (int) hostEntity.getPosY()), p, false, PathFinding.CornerInclusionRule.NON_CLIPPING_CORNERS, 0, true);
                 if (currentPath == null) {
                     targetUnreachable = true;
@@ -70,14 +70,14 @@ public class ChaseAI extends AIState {
                 }
             } else {
                 Point nextNode = currentPath.get(0);
-                float targetX = nextNode.x + 0.5f;
-                float targetY = nextNode.y + 0.5f;
-                float distToNextTile = WorldUtils.getRawDistance(hostEntity.getPosX(), targetX, hostEntity.getPosY(), targetY);
-                float angleToNextTile = (float) Math.atan2(targetY - hostEntity.getPosY(), targetX - hostEntity.getPosX());
+                double targetX = nextNode.x + 0.5f;
+                double targetY = nextNode.y + 0.5f;
+                double distToNextTile = WorldUtils.getRawDistance(hostEntity.getPosX(), targetX, hostEntity.getPosY(), targetY);
+                double angleToNextTile = (double) Math.atan2(targetY - hostEntity.getPosY(), targetX - hostEntity.getPosX());
                 if (distToNextTile > WorldUtils.ENTITY_WITHIN_TILE_THRESHOLD) {
-                    float speed = hostEntity.getWalkSpeed() * Math.min(1, distToNextTile * 2);
-                    float velX = speed * (float) Math.cos(angleToNextTile);
-                    float velY = speed * (float) Math.sin(angleToNextTile);
+                    double speed = hostEntity.getWalkSpeed() * Math.min(1, distToNextTile * 2);
+                    double velX = speed * (double) Math.cos(angleToNextTile);
+                    double velY = speed * (double) Math.sin(angleToNextTile);
                     hostEntity.addVelocity(velX, velY);
                 } else {
                     currentPath.remove(0);
@@ -117,16 +117,16 @@ public class ChaseAI extends AIState {
     }
 
     /**
-     * Generate a map of float weights for pathfinding @PathFinding.Java
+     * Generate a map of double weights for pathfinding @PathFinding.Java
      */
-    public float[][] getWeightMapFromMemory() {
+    public double[][] getWeightMapFromMemory() {
 
-        float[][] weightMap = new float[hostEntity.getFloor().getHeight()][hostEntity.getFloor().getWidth()];
+        double[][] weightMap = new double[hostEntity.getFloor().getHeight()][hostEntity.getFloor().getWidth()];
 
         for (int y = 0; y < hostEntity.getFloor().getHeight(); y++) {
             for (int x = 0; x < hostEntity.getFloor().getWidth(); x++) {
 
-                float weight;
+                double weight;
                 CellMemory mem = hostEntity.getCurrentFloorMemory().getDataAt(x, y);
 
                 if (mem == null || mem.enterable == CellMemory.EnterableStatus.CLOSED) {
